requirements.txt
pandas==2.0.3
seaborn==0.12.2
matplotlib==3.7.2
scikit-learn==1.3.2
kafka-python==2.0.2